+++ common/src/main/java/org/commonjava/storage/pathmapped/model/FileChecksum.java
