+++ common/src/main/java/org/commonjava/storage/pathmapped/model/PathMap.java
