+++ pathdb/pom.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>path-mapped</artifactId>
+        <groupId>org.commonjava.util</groupId>
+        <version>0.1-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>path-mapped-pathdb</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>datastax</module>
+        <module>jpa</module>
+    </modules>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.commonjava.util</groupId>
+            <artifactId>path-mapped-common</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
