+++ storage/pom.xml
@@ -12,42 +12,20 @@
     <artifactId>path-mapped-storage</artifactId>
 
     <dependencies>
-        <!-- JPA support-->
         <dependency>
-            <groupId>javax.persistence</groupId>
-            <artifactId>javax.persistence-api</artifactId>
-            <version>2.2</version>
-        </dependency>
-
-        <!--Datastax Java Driver-->
-        <dependency>
-            <groupId>com.datastax.cassandra</groupId>
-            <artifactId>cassandra-driver-core</artifactId>
-            <version>${datastaxVersion}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.datastax.cassandra</groupId>
-            <artifactId>cassandra-driver-mapping</artifactId>
-            <version>${datastaxVersion}</version>
+            <groupId>org.commonjava.util</groupId>
+            <artifactId>path-mapped-common</artifactId>
         </dependency>
 
         <!--Test-->
         <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-entitymanager</artifactId>
-            <version>${hibernateVersion}</version>
+            <groupId>org.commonjava.util</groupId>
+            <artifactId>path-mapped-pathdb-datastax</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-core</artifactId>
-            <version>${hibernateVersion}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>${h2Version}</version>
+            <groupId>org.commonjava.util</groupId>
+            <artifactId>path-mapped-pathdb-jpa</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
