+++ storage/src/test/java/org/commonjava/storage/pathmapped/PathMapUtilsTest.java
@@ -1,17 +1,15 @@
 package org.commonjava.storage.pathmapped;
 
-import org.commonjava.storage.pathmapped.jpa.model.JpaPathKey;
-import org.commonjava.storage.pathmapped.jpa.model.JpaPathMap;
+import org.commonjava.storage.pathmapped.pathdb.jpa.model.JpaPathKey;
+import org.commonjava.storage.pathmapped.pathdb.jpa.model.JpaPathMap;
+import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.List;
 
 import static org.commonjava.storage.pathmapped.util.PathMapUtils.getFilename;
 import static org.commonjava.storage.pathmapped.util.PathMapUtils.getParentPath;
-
 import static org.commonjava.storage.pathmapped.util.PathMapUtils.getParents;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
 
 public class PathMapUtilsTest
 {
@@ -32,39 +30,39 @@ public class PathMapUtilsTest
         {
             System.out.println( ">> " + map.getPathKey() );
         }
-        assertEquals( 3, l.size() );
+        Assert.assertEquals( 3, l.size() );
     }
 
     @Test
     public void getParentPathTest()
     {
         String parent = getParentPath( "/path/to/file.txt" );
-        assertEquals( parent, "/path/to" );
+        Assert.assertEquals( parent, "/path/to" );
 
         parent = getParentPath( "/path/to/" );
-        assertEquals( parent, "/path" );
+        Assert.assertEquals( parent, "/path" );
 
         parent = getParentPath( "/path" );
-        assertEquals( parent, "/" );
+        Assert.assertEquals( parent, "/" );
 
         parent = getParentPath( "/" );
-        assertNull( parent );
+        Assert.assertNull( parent );
     }
 
     @Test
     public void getFilenameTest()
     {
         String filename = getFilename( "/path/to/file.txt" );
-        assertEquals( filename, "file.txt" );
+        Assert.assertEquals( filename, "file.txt" );
 
         filename = getFilename( "/path/to/" );
-        assertEquals( filename, "to/" );
+        Assert.assertEquals( filename, "to/" );
 
         filename = getFilename( "/path/" );
-        assertEquals( filename, "path/" );
+        Assert.assertEquals( filename, "path/" );
 
         filename = getFilename( "/" );
-        assertNull( filename );
+        Assert.assertNull( filename );
     }
 
 }
