+++ tool/pom.xml
@@ -16,10 +16,13 @@
     </properties>
 
     <dependencies>
-
         <dependency>
             <groupId>org.commonjava.util</groupId>
             <artifactId>path-mapped-storage</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.commonjava.util</groupId>
+            <artifactId>path-mapped-pathdb-datastax</artifactId>
         </dependency>
 
         <!-- Tool -->
